+++ src/app/custom-content/has-custom-content.directive.spec.ts
@@ -1,17 +1,16 @@
-import { NgIf } from "@angular/common";
-import { Component, Directive, inject, Input } from "@angular/core";
+import { Component } from "@angular/core";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { By } from "@angular/platform-browser";
 import { BehaviorSubject, of } from "rxjs";
-import { CustomContent, CustomContentService } from "./custom-content.service";
-import { HasContentForSlotDirective } from "./has-content-for-slot.directive";
+import { CustomContentService, SlotContent } from "./custom-content.service";
+import { HasCustomContentDirective } from "./has-custom-content.directive";
 
-describe("HasContentForSlotDirective", () => {
+describe("HasCustomContentDirective", () => {
   let fixture: ComponentFixture<TestHostComponent>;
   beforeEach(async() => {
     TestBed.configureTestingModule({
       declarations: [TestHostComponent],
-      imports:[HasContentForSlotDirective],
+      imports:[HasCustomContentDirective],
       providers: [
         {provide:CustomContentService, useClass:CustomContentServiceStub}
       ]
@@ -39,34 +38,16 @@ describe("HasContentForSlotDirective", () => {
 
 
 @Component({
-  template: `<div *appHasContentForSlot="'test-slot'" class="test-content">Test Element</div>`
+  template: `<div *appHasCustomContent="'test-slot'" class="test-content">Test Element</div>`
 })
 class TestHostComponent {}
 
 class CustomContentServiceStub {
   private _contents$ = new BehaviorSubject([]);
-  contents$ = this._contents$.asObservable();
-  contentLoaded$ = of(true);
+  customContent$ = this._contents$.asObservable();
+  customContentLoaded$ = of(true);
 
-  setContent(c: CustomContent[]) {
+  setContent(c: SlotContent[]) {
     this._contents$.next(c);
-  }
-}
-
-//mock for other components that will use the directive
-@Directive({
-  selector: '[appHasContentForSlot]',
-  hostDirectives: [{
-     directive: NgIf,
-     inputs: ['ngIfElse: appHasContentForSlotElse'] },
-  ],
-  standalone: true,
-})
-export class HasContentForSlotMockDirective {
-  @Input("appHasContentForSlot") slotId: string;
-  private ngIfDirective = inject(NgIf);
-
-  constructor() {
-    this.ngIfDirective.ngIf = true;
   }
 }